+++ src/components/form/MultivalueField/multivalueField.js
@@ -4,7 +4,7 @@ import './multivalueField.scss';
 import { domContentLoadedWrapper } from '../../../domContentLoadedWrapper';
 
 export default function () {
-  function addField(scope) {
+  function addField(scope, initialValue = '') {
     const wrappers = scope.querySelectorAll('.input-field__wrapper');
     const wrapper = wrappers[wrappers.length - 1];
 
@@ -14,7 +14,7 @@ export default function () {
     if (clonedDeleteButton) {
       clonedDeleteButton.remove();
     }
-    clonedEl.querySelector('input').value = '';
+    clonedEl.querySelector('input').value = initialValue;
 
     const deleteButton = document.createElement('div');
     deleteButton.classList.add('multivalue-field__delete-entity');
@@ -34,6 +34,12 @@ export default function () {
     const multivalueFields = document.querySelectorAll('.multivalue-field');
 
     for (const field of multivalueFields) {
+      const values = JSON.parse(field.querySelector('.multivalue-field__values').innerHTML).slice(1);
+
+      for (const value of values) {
+        addField(field, value);
+      }
+
       field.querySelector('.multivalue-field__add-entity').addEventListener('click', function (e) {
         e.preventDefault();
         addField(field);