+++ src/index.js
@@ -55,6 +55,7 @@ import './components/form/Step/step';
 import mapField from './components/form/MapField/mapField';
 import tooltip from './components/form/Tooltip/tooltip';
 import multivalueField from './components/form/MultivalueField/multivalueField';
+import multifieldGroup from './components/form/MultifieldGroup/multifieldGroup';
 
 /* Landing Components */
 import './components/landing/LandingLinks/landingLinks';
@@ -152,3 +153,4 @@ dropdownMenu();
 webtoolProgressBar();
 tooltip();
 multivalueField();
+multifieldGroup();